third_party/blink/web_tests/wpt_internal/css/css-display/visibility-inert-click.html

<!DOCTYPE html>
<title>CSS Display Test: visibility:inert blocks click events</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/10711#issuecomment-2338834244">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  #inert {
    visibility: inert;
    width: 100px;
    height: 100px;
    background: lime;
  }
</style>
<div id="inert"></div>
<script>
  let clicked = false;
  inert.addEventListener('click', () => clicked = true);

  promise_test(async () => {
    try {
      await test_driver.click(inert);
      assert_false(clicked);
    } catch (e) {
      // test driver detects inert elements as unclickable and throws an error
      assert_false(clicked);
    }
  }, "No click event for visibility:inert element");

  promise_test(async () => {
    inert.style.visibility = 'visible';
    await test_driver.click(inert);
    assert_true(clicked);
  }, "Click event should work after visibility change");
</script>
